Stabilize `boilercv.morphs.ContextMorph`. Implement subcooling correlations
