+++ changelog/188.change.md
@@ -1,1 +1,1 @@
-Implement subcooling correlations
+Stabilize `boilercv.morphs.ContextMorph`. Implement subcooling correlations
