+++ pipeline/boilercv_pipeline/correlations/dimensionless_bubble_diameter/morphs.py
@@ -7,7 +7,9 @@ from numpy import linspace, pi
 from sympy import symbols
 from tomlkit import parse
 
-from boilercv.morphs import Morph
+from boilercv.mappings.models import Repl
+from boilercv.morphs.contexts import LocalSymbols, compose_context
+from boilercv.morphs.morphs import Morph
 from boilercv_pipeline.correlations.annotations import Kind
 from boilercv_pipeline.correlations.dimensionless_bubble_diameter.types import (
     Param,
@@ -16,8 +18,6 @@ from boilercv_pipeline.correlations.dimensionless_bubble_diameter.types import (
     syms,
 )
 from boilercv_pipeline.correlations.models import Equations, EquationSolutions
-from boilercv_pipeline.mappings.models import Repl
-from boilercv_pipeline.morphs import LocalSymbols, compose_context
 from boilercv_pipeline.types import Expectation
 from boilercv_pipeline.types.runtime import SympifyParams
