--- pipeline/boilercv_pipeline/mappings/types.py
@@ -1,6 +1,0 @@
-"""Types for static typing of mappings."""
-
-from typing import TypeVar
-
-K = TypeVar("K")
-T = TypeVar("T")
