+++ tests/boilercv_tests/test_morphs.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Any, Generic, Literal, TypeAlias
 import pytest
 from pydantic import ValidationError
 
-from boilercv.morphs import K, Morph, TypeType, V
+from boilercv.morphs.morphs import K, Morph, TypeType, V
 
 # fmt: off
